Update: calculateUnits & stockStateToCSV
- Moved calculateUnits and its helper functions into InventoryController
- Updated calculateUnits to check invIDs before checking the rest of the
  supply list
- Changed calculateUnits to calculateUnitsAndStates
- Added an API route for calculateUnitsAndStates
- Added a temporary test method that will be removed later

- Added calculatedMinQuantity and calculatedStockState to Stock Report
  XLSX generation
- Moved comment for filling description row above the cleanUpCSV call
  for item.description
