+++ server/src/main/java/umm3601/Inventory/InventoryController.java
@@ -8,6 +8,7 @@ import static com.mongodb.client.model.Filters.regex;
 
 // Java Imports
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.regex.Pattern;
@@ -38,7 +39,9 @@ import io.javalin.http.NotFoundResponse;
 import umm3601.Auth.HttpMethod;
 import umm3601.Auth.RequirePermission;
 import umm3601.Auth.Route;
+import umm3601.Family.Family;
 import umm3601.Family.InventoryReservationService;
+import umm3601.SupplyList.SupplyList;
 
 
 // Controller
@@ -49,6 +52,7 @@ public class InventoryController {
   private static final String API_INVENTORY_REMOVE_QUANTITY = "/api/inventory/removeQuantity";
   private static final String API_INVENTORY_CLEAR = "/api/inventory/clear";
   private static final String API_INVENTORY_RESET = "/api/inventory/resetQuantity";
+  private static final String API_CALCULATE_STATES = "/api/inventory/calculateStates";
 
   static final String ITEM_KEY = "item";
   static final String BRAND_KEY = "brand";
@@ -74,6 +78,9 @@ public class InventoryController {
   private final InventoryReservationService inventoryReservationService;
   private final InventoryIdService inventoryIdService;
 
+  private final JacksonMongoCollection<Family> familyCollection;
+  private final JacksonMongoCollection<SupplyList> supplyListCollection;
+
   public InventoryController(MongoDatabase database) {
     this(database, new InventoryReservationService(database), new InventoryIdService(database));
   }
@@ -95,6 +102,20 @@ public class InventoryController {
       Inventory.class,
       UuidRepresentation.STANDARD
     );
+
+    familyCollection = JacksonMongoCollection.builder().build(
+      database,
+      "family",
+      Family.class,
+      UuidRepresentation.STANDARD
+    );
+
+    supplyListCollection = JacksonMongoCollection.builder().build(
+      database,
+      "supplyList",
+      SupplyList.class,
+      UuidRepresentation.STANDARD
+    );
   }
 
   // Endpoint to generate the next internal ID
@@ -110,7 +131,7 @@ public class InventoryController {
    * @param ctx The context for the HTTP request
    */
   @Route(method = HttpMethod.POST, path = API_INVENTORY)
-  @RequirePermission("add_inventory_item")
+  @RequirePermission("edit_inventory_item")
   public void addInventory(Context ctx) {
     Inventory newInv = ctx.bodyAsClass(Inventory.class);
     newInv.refreshDescription();
@@ -519,5 +540,175 @@ public class InventoryController {
       }
       inventoryCollection.updateOne(eq("_id", new ObjectId(inv._id)), Updates.set("stockState", inv.stockState));
     }
+  }
+
+  /**
+   * Gets the total number of students in each school and grade.
+   * @return a map of school to grade to teacher to student count
+   */
+  private Map<String, Map<String, Map<String, Integer>>> getSchoolGradeTeacherTotals() {
+    ArrayList<Family> families = familyCollection.find().into(new ArrayList<>());
+    Map<String, Map<String, Map<String, Integer>>> schoolGradeTeacherTotals = new HashMap<>();
+
+    // Go through all the families
+    for (Family family : families) {
+      if (family == null || family.students == null) {
+        continue;
+      }
+
+      // Go through all the students in that family
+      for (Family.StudentInfo student : family.students) {
+        if (student == null) {
+          continue;
+        }
+
+        // Get the school, grade, and teacher for the student
+        String school = student.school != null ? student.school : "Unknown School";
+        String grade = student.grade != null ? student.grade : "Unknown Grade";
+        String teacher = student.teacher != null ? student.teacher : "Unknown Teacher";
+
+        // Add the student to the appropriate school
+        Map<String, Map<String, Integer>> gradeTotals = schoolGradeTeacherTotals.get(school);
+        if (gradeTotals == null) {
+          gradeTotals = new HashMap<>();
+          schoolGradeTeacherTotals.put(school, gradeTotals); // Put grades in schools
+        }
+
+        Map<String, Integer> teacherTotals = gradeTotals.get(grade);
+        if (teacherTotals == null) {
+          teacherTotals = new HashMap<>();
+          gradeTotals.put(grade, teacherTotals); // Put teachers in grades
+        }
+
+        teacherTotals.put(teacher, teacherTotals.getOrDefault(teacher, 0) + 1);  // Add student to teacher
+      }
+    }
+
+    return schoolGradeTeacherTotals;
+  }
+
+  /**
+   * Updates the calculatedStockState of the given inventory item based on quantity, minQuantity, and maxQuantity.
+   * Use this method to update the calculatedStockState of an inventory item whenever its
+   * calculatedMinQuantity changes.
+   *
+   * @param inv The inventory item to update the calculatedStockState for
+   * @throws NotFoundResponse if the item was not found
+   * @throws IllegalArgumentException if calculatedMinQuantity is null
+   * @return the updated calculatedStockState of the inventory item
+  */
+  private String calculateStockState(Inventory inv) {
+    String calculatedStockState = null;
+    // Make sure item exists
+    if (inv == null) {
+      throw new NotFoundResponse("The requested inventory item was not found");
+    } else {
+      // Validate quantity, minQuantity, and maxQuantity
+      if (inv.calculatedMinQuantity == null) {
+        throw new IllegalArgumentException("calculatedMinQuantity must be an integer.");
+      }
+
+      int itemDiff = inv.quantity - inv.calculatedMinQuantity;
+
+      if (itemDiff == 0) {
+        calculatedStockState = "Stocked";
+      } else if (itemDiff < 0) {
+        calculatedStockState = "Understocked";
+      } else if (itemDiff > 0) {
+        calculatedStockState = "Overstocked";
+      } else {
+        calculatedStockState = "Unknown";
+      }
+
+      return calculatedStockState;
+    }
+  }
+
+  /**
+   * Calculates the calculatedStockState and calculatedMinQuantity of items based on number of students under each teacher, grade, and school.
+   */
+  private void calculatesUnitsAndStates(){
+    ArrayList<SupplyList> allSupplyLists = supplyListCollection.find().into(new ArrayList<>());
+
+    // loop through each supply list
+    for (SupplyList supplyList : allSupplyLists) {
+      int totalNeeded = 0;
+
+      if (supplyList.school == null || supplyList.grade == null || supplyList.teacher == null) {
+        continue; // Skip if any of the fields are null
+      }
+
+      // Find first properly formatted invID
+      String validInvID = null;
+      for (String invID : supplyList.invIDs) {
+        if (invID != null && invID.matches("^ID-\\d{5}$")) {
+          validInvID = invID;
+          break;
+        }
+      }
+
+      if (validInvID == null) { // No properly formatted ID found
+        supplyList.percentageFilled = -2;
+        supplyListCollection.updateOne(eq("_id", new ObjectId(supplyList._id)), Updates.set("percentageFilled", supplyList.percentageFilled));
+        continue; // Skip to next supply list
+      }
+
+      Inventory bestMatch = inventoryCollection.find(eq("internalID", validInvID)).first();
+
+      if (bestMatch == null) {
+        continue; // Skip if no matching inventory item is found
+      }
+
+      // loop through each school
+      for (Map.Entry<String, Map<String, Map<String, Integer>>> schoolEntry : getSchoolGradeTeacherTotals().entrySet()) {
+        String school = schoolEntry.getKey();
+        Map<String, Map<String, Integer>> gradeTotals = schoolEntry.getValue();
+
+        // loop through each grade
+        for (Map.Entry<String, Map<String, Integer>> gradeEntry : gradeTotals.entrySet()) {
+          String grade = gradeEntry.getKey();
+          Map<String, Integer> teacherTotals = gradeEntry.getValue();
+
+          // loop through each teacher
+          for (Map.Entry<String, Integer> teacherEntry : teacherTotals.entrySet()) {
+            String teacher = teacherEntry.getKey();
+            int numStudents = teacherEntry.getValue();
+
+            // if the supply list matches the school, grade, and teacher,
+            // add the quantity needed for that supply list to the total needed
+            if (supplyList.school.equals(school) && (supplyList.grade.equals(grade)) && (supplyList.teacher.equals(teacher))) {
+              int qty = supplyList.quantity != null ? supplyList.quantity : 1;
+              totalNeeded += numStudents * qty;
+
+              // Use the first linked item in the supply list to find the corresponding inventory item and update its calculatedMinQuantity
+
+              if (bestMatch != null) {
+                bestMatch.calculatedMinQuantity = totalNeeded;
+                bestMatch.calculatedStockState = calculateStockState(bestMatch);
+                inventoryCollection.updateOne(eq("_id", new ObjectId(bestMatch._id)), Updates.set("calculatedMinQuantity", bestMatch.calculatedMinQuantity));
+                inventoryCollection.updateOne(eq("_id", new ObjectId(bestMatch._id)), Updates.set("calculatedStockState", bestMatch.calculatedStockState));
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+
+  private void test(){
+    FindIterable<Inventory> results = inventoryCollection.find();
+    ArrayList<Inventory> matching = results.into(new ArrayList<>());
+    for (Inventory inv : matching) {
+      calculateStockState(inv);
+    }
+  }
+
+  // Endpoint to calculate states
+  @Route(method = HttpMethod.GET, path = API_CALCULATE_STATES)
+  @RequirePermission("add_inventory_item")
+  public void calculateStates(Context ctx) {
+    test();
+    ctx.json(Map.of("message", "Calculated states successfully"));
+    ctx.status(HttpStatus.OK);
   }
 }
